Fix initialize.py
